+++ include/response.h
@@ -39,7 +39,7 @@ typedef struct Response {
 
     // all about the order this response is about
     u32 order_id;// this might be in response to an order, set to U32MAX if this is broadcast
-    u16 status;
+    u32 status;
     u32 quantity_filled;
     u16 price;
